Add files via upload
Hello,

This was part of a test given to me by a company.  They wanted a solution and they got it right here.

The test:   
Create a railroad that will help passengers with information about routes.  You are supposed to show outputs such as the distance around towns, and what is the shortest route for two towns in between. 

It involves a railroad train service that many towns use in Kiwiland.  All the tracks are "one-way". 

Keep in mind that an edge represents a route between two houses, and a node represents a town.  Furthermore, Output: For test input 1 through 5, if no such route exists, output 'NO SUCH ROUTE'. 

You must provide the following:

1. The distance of the route A-B-C.
2. The distance of the route A-D.
3. The distance of the route A-D-C.
4. The distance of the route A-E-B-C-D.
5. The distance of the route A-E-D.
6. The number of trips starting at C and ending at C with a maximum of 3
stops.  In the sample data below, there are two such trips: C-D-C (2
stops). and C-E-B-C (3 stops).
7. The number of trips starting at A and ending at C with exactly 4 stops.
In the sample data below, there are three such trips: A to C (via B,C,D); A
to C (via D,C,D); and A to C (via D,E,B).
8. The length of the shortest route (in terms of distance to travel) from A
to C.
9. The length of the shortest route (in terms of distance to travel) from B
to B.
10. The number of different routes from C to C with a distance of less than 30.
In the sample data, the trips are: CDC, CEBC, CEBCDC, CDCEBC, CDEBC,
CEBCEBC, CEBCEBCEBC.

Test Input:

For the test input, the towns are named using the first few letters of the alphabet from A to D.  A route between two towns (A to B) with a distance of 5 is represented as AB5.

Graph: AB5, BC4, CD8, DC8, DE6, AD5, CE2, EB3, AE7